Remove and reuse connection class
