+++ src/net/NetworkServer.cpp
@@ -1,55 +1,5 @@
 #include "NetworkServer.h"
 
-tcp::socket& Connection::get_socket() {
-    return socket;
-}
-
-void Connection::start_async_read_loop() {
-    socket.async_read_some(boost::asio::buffer(rcvbuf),
-        [&](const boost::system::error_code& error, size_t n_bytes) {
-            if (!error) {
-                message_queue.push(rcvbuf);
-                std::cerr << "[s] recv: " << string(rcvbuf);
-            }
-            else if (error != boost::asio::error::eof) {
-                std::cerr << "FATAL handle_read error: " << error << "\n";
-                return;
-            }
-
-            // Wait for and read the next message.
-            start_async_read_loop();
-        });
-}
-
-bool Connection::send(const string& message) {
-    if (!socket.is_open()) {
-        return false;
-    }
-
-    boost::system::error_code error;
-    boost::asio::write(socket, boost::asio::buffer(message), error);
-
-    // If we get these errors, it's likely a clean disconnect.
-    if ((error == boost::asio::error::eof) ||
-        (error == boost::asio::error::connection_reset) ||
-        (error == boost::asio::error::broken_pipe)) {
-        std::cerr << "[send] client disconnected.\n";
-        return false;
-    } else if (error) {
-        std::cerr << "[send] some other error: " << error << "\n";
-        return false;
-    }
-
-    return true;
-}
-
-bool Connection::read_message(string* message) {
-    if (message_queue.empty())
-        return false;
-    *message = message_queue.pop();
-    return true;
-}
-
 NetworkServer::NetworkServer(boost::asio::io_service& ios, unsigned int port) :
     acceptor(ios, tcp::endpoint(tcp::v4(), port)), next_id(0) {
     start_accept();
